% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_plate_barcodes.R
\name{split_plate_barcodes}
\alias{split_plate_barcodes}
\title{Split by plate barcodes}
\usage{
split_plate_barcodes(
  pair1_filepath,
  pair2_filepath,
  forward_barcode,
  nreads = 100,
  output_prefix = "outputfile"
)
}
\arguments{
\item{pair1_filepath}{file path of the first paired end file}

\item{pair2_filepath}{file path of the second paired end file}

\item{forward_barcode}{A character vector with the foward barcodes}

\item{nreads}{number of reads to iterate by}

\item{output_prefix}{The output prefix for the output files}
}
\description{
Split fastq files by the plate barcodes provided
}
